+++ src/Core/EBF_AnalogInput.h
@@ -21,7 +21,7 @@ class EBF_AnalogInput : protected EBF_HalInstance {
 		float GetLastValue();
 		uint16_t GetRawLastValue();
 
-		void SetPollInterval(uint16_t ms);
+		void SetPollInterval(uint32_t ms);
 
 	protected:
 		uint8_t Process();
