+++ src/Core/EBF_DigitalInput.cpp
@@ -53,7 +53,7 @@ uint8_t EBF_DigitalInput::Init(
 	return EBF_OK;
 }
 
-void EBF_DigitalInput::SetPollInterval(uint16_t ms)
+void EBF_DigitalInput::SetPollInterval(uint32_t ms)
 {
 	// No polling needed if there is no callback to call
 	if (callbackFunc == NULL) {
