+++ src/Core/EBF_I2C.cpp
@@ -25,7 +25,7 @@ uint8_t EBF_I2C::Init(EBF_CallbackType callbackFunc, uint8_t address)
 	return EBF_OK;
 }
 
-void EBF_I2C::SetPollInterval(uint16_t ms)
+void EBF_I2C::SetPollInterval(uint32_t ms)
 {
 	// No polling needed if there is no callback to call
 	if (callbackFunc == NULL) {
