+++ src/Core/EBF_Serial.cpp
@@ -68,7 +68,7 @@ uint8_t EBF_Serial::Init(
 	return EBF_OK;
 }
 
-void EBF_Serial::SetPollInterval(uint16_t ms)
+void EBF_Serial::SetPollInterval(uint32_t ms)
 {
 	// No polling needed if there is no callback to call
 	if (callbackFunc == NULL) {
