+++ src/Core/EBF_Timers.h
@@ -14,8 +14,8 @@ class EBF_Timers {
 		uint8_t StartTimer(uint8_t timerId, uint16_t milliSec);
 		uint8_t StopTimer(uint8_t timerId);
 
-		// Returns number of milliSeconds till the next timer, or 0xFFFF if there are no timers
-		uint16_t Process(unsigned long current);
+		// Returns number of milliSeconds till the next timer, or 0xFFFFFFFF if there are no timers
+		uint32_t Process(unsigned long current);
 
 	private:
 		typedef struct
